project/recommendation_service/requirements.txt
django==4.1.3
djangorestframework==3.14.0
psycopg2==2.9.5
jsonfield==3.1.0